src/components/shared/LoadingPlaceholder/LoadingPlaceholder.spec.tsx
import { render } from '@testing-library/react';
import LoadingPlaceholder, {
  LoadingPlaceholderCircleSize,
  LoadingPlaceholderHeight,
  LoadingPlaceholderProps,
  LoadingPlaceholderWidth,
} from './LoadingPlaceholder';

const subject = (props: LoadingPlaceholderProps) => render(<LoadingPlaceholder {...props} />);

describe('LoadingPlaceholder', () => {
  ['tiny', 'small', 'medium', 'large'].forEach(width => {
    ['small', 'medium'].forEach(height => {
      it(`renders a loading placeholder with ${width} width and ${height} height`, () => {
        const { container } = subject({
          shape: {
            width: width as LoadingPlaceholderWidth,
            height: height as LoadingPlaceholderHeight,
          },
        });

        expect(container).not.toBeNull();
        expect(container).toMatchSnapshot();
      });
    });
  });

  ['small', 'medium'].forEach(size => {
    it(`renders a circular loading placeholder with ${size} size`, () => {
      const { container } = subject({
        shape: {
          circle: size as LoadingPlaceholderCircleSize,
        },
      });

      expect(container).not.toBeNull();
      expect(container).toMatchSnapshot();
    });
  });
});
